+++ src/code42cli/cmds/alerts.py
@@ -3,7 +3,6 @@ from _collections import OrderedDict
 import click
 import py42.sdk.queries.alerts.filters as f
 from c42eventextractor.extractors import AlertExtractor
-from click import echo
 
 import code42cli.cmds.search.enums as enum
 import code42cli.cmds.search.extraction as ext
@@ -11,8 +10,13 @@ import code42cli.cmds.search.options as searchopt
 import code42cli.errors as errors
 import code42cli.options as opt
 from code42cli.cmds.search.cursor_store import AlertCursorStore
+from code42cli.cmds.search.extraction import handle_no_events
+from code42cli.logger import get_logger_for_server
 from code42cli.options import format_option
+from code42cli.options import server_options
+from code42cli.output_formats import JsonOutputFormat
 from code42cli.output_formats import OutputFormatter
+
 
 SEARCH_DEFAULT_HEADER = OrderedDict()
 SEARCH_DEFAULT_HEADER["name"] = "RuleName"
@@ -21,6 +25,7 @@ SEARCH_DEFAULT_HEADER["createdAt"] = "ObservedDate"
 SEARCH_DEFAULT_HEADER["state"] = "Status"
 SEARCH_DEFAULT_HEADER["severity"] = "Severity"
 SEARCH_DEFAULT_HEADER["description"] = "Description"
+
 
 search_options = searchopt.create_search_options("alerts")
 
@@ -122,6 +127,14 @@ description_option = click.option(
     help="Filter alerts by description. Does fuzzy search by default.",
 )
 
+send_to_format_options = click.option(
+    "-f",
+    "--format",
+    type=click.Choice(JsonOutputFormat(), case_sensitive=False),
+    help="The output format of the result. Defaults to json format.",
+    default=JsonOutputFormat.JSON,
+)
+
 
 def alert_options(f):
     f = actor_option(f)
@@ -137,7 +150,6 @@ def alert_options(f):
     f = description_option(f)
     f = severity_option(f)
     f = state_option(f)
-    f = format_option(f)
     return f
 
 
@@ -157,6 +169,21 @@ def clear_checkpoint(state, checkpoint_name):
     _get_alert_cursor_store(state.profile.name).delete(checkpoint_name)
 
 
+def _call_extractor(
+    cli_state, handlers, begin, end, or_query, advanced_query, **kwargs
+):
+    extractor = _get_alert_extractor(cli_state.sdk, handlers)
+    extractor.use_or_query = or_query
+    if advanced_query:
+        extractor.extract_advanced(advanced_query)
+    else:
+        if begin or end:
+            cli_state.search_filters.append(
+                ext.create_time_range_filter(f.DateObserved, begin, end)
+            )
+        extractor.extract(*cli_state.search_filters)
+
+
 @alerts.command()
 @alert_options
 @search_options
@@ -170,6 +197,7 @@ def clear_checkpoint(state, checkpoint_name):
     is_flag=True,
     help="Display simple properties of the primary level of the nested response.",
 )
+@format_option
 def search(
     cli_state,
     format,
@@ -195,18 +223,45 @@ def search(
         formatter=formatter,
         force_pager=include_all,
     )
-    extractor = _get_alert_extractor(cli_state.sdk, handlers)
-    extractor.use_or_query = or_query
-    if advanced_query:
-        extractor.extract_advanced(advanced_query)
-    else:
-        if begin or end:
-            cli_state.search_filters.append(
-                ext.create_time_range_filter(f.DateObserved, begin, end)
-            )
-        extractor.extract(*cli_state.search_filters)
-    if handlers.TOTAL_EVENTS == 0 and not errors.ERRORED:
-        echo("No results found.")
+    _call_extractor(cli_state, handlers, begin, end, or_query, advanced_query, **kwargs)
+    handle_no_events(not handlers.TOTAL_EVENTS and not errors.ERRORED)
+
+
+@alerts.command()
+@alert_options
+@search_options
+@click.option(
+    "--or-query", is_flag=True, cls=searchopt.AdvancedQueryAndSavedSearchIncompatible
+)
+@opt.sdk_options()
+@server_options
+@click.option(
+    "--include-all",
+    default=False,
+    is_flag=True,
+    help="Display simple properties of the primary level of the nested response.",
+)
+@send_to_format_options
+def send_to(
+    cli_state,
+    format,
+    hostname,
+    protocol,
+    begin,
+    end,
+    advanced_query,
+    use_checkpoint,
+    or_query,
+    **kwargs
+):
+    """Send alerts to the given server address."""
+    logger = get_logger_for_server(hostname, protocol, format)
+    cursor = _get_alert_cursor_store(cli_state.profile.name) if use_checkpoint else None
+    handlers = ext.create_send_to_handlers(
+        cli_state.sdk, AlertExtractor, cursor, use_checkpoint, logger,
+    )
+    _call_extractor(cli_state, handlers, begin, end, or_query, advanced_query, **kwargs)
+    handle_no_events(not handlers.TOTAL_EVENTS and not errors.ERRORED)
 
 
 def _get_alert_extractor(sdk, handlers):