+++ src/code42cli/cmds/securitydata_output_formats.py
@@ -37,7 +37,6 @@ def _convert_event_to_cef(event):
     }
 
     extension = " ".join(_format_cef_kvp(key, kvp_list[key]) for key in kvp_list)
-
     event_name = event.get("eventType", "UNKNOWN")
     signature_id = FILE_EVENT_TO_SIGNATURE_ID_MAP.get(event_name, "C42000")