+++ src/code42cli/util.py
@@ -147,6 +147,14 @@ class warn_interrupt:
         return inner
 
 
+def get_url_parts(url_str):
+    parts = url_str.split(":")
+    port = None
+    if len(parts) > 1 and parts[1] != "":
+        port = int(parts[1])
+    return parts[0], port
+
+
 def _get_default_header(header_items):
     if not header_items:
         return