+++ tests/cmds/conftest.py
@@ -35,6 +35,15 @@ def cli_logger(mocker):
 
 
 @pytest.fixture
+def event_extractor_logger(mocker):
+    mock = mocker.patch(
+        "c42eventextractor.logging.handlers.NoPrioritySysLogHandlerWrapper"
+    )
+    mock.emit.return_value = mocker.MagicMock()
+    return mock
+
+
+@pytest.fixture
 def cli_state_with_user(sdk_with_user, cli_state):
     cli_state.sdk = sdk_with_user
     return cli_state