+++ tests/cmds/search/test_extraction.py
@@ -6,8 +6,10 @@ from requests import Response
 from code42cli import errors
 from code42cli.cmds.search.cursor_store import BaseCursorStore
 from code42cli.cmds.search.extraction import create_handlers
+from code42cli.cmds.search.extraction import create_send_to_handlers
 from code42cli.cmds.search.extraction import try_get_default_header
 from code42cli.output_formats import OutputFormat
+
 
 key = "events"
 
@@ -63,3 +65,26 @@ def test_create_handlers_creates_handlers_that_pass_events_to_output_format(
     py42_response = Py42Response(http_response)
     handlers.handle_response(py42_response)
     formatter.get_formatted_output.assert_called_once_with(events)
+
+
+def test_send_to_handlers_creates_handlers_that_pass_events_to_logger(
+    mocker, sdk, event_extractor_logger
+):
+    class TestExtractor(BaseExtractor):
+        def __init__(self, handlers, timestamp_filter):
+            timestamp_filter._term = "test_term"
+            super().__init__(key, search, handlers, timestamp_filter, TestQuery)
+
+        def _get_timestamp_from_item(self, item):
+            pass
+
+    cursor_store = mocker.MagicMock(sepc=BaseCursorStore)
+    handlers = create_send_to_handlers(
+        sdk, TestExtractor, cursor_store, "chk-name", event_extractor_logger
+    )
+    http_response = mocker.MagicMock(spec=Response)
+    events = [{"property": "bar"}]
+    http_response.text = '{{"{0}": [{{"property": "bar"}}]}}'.format(key)
+    py42_response = Py42Response(http_response)
+    handlers.handle_response(py42_response)
+    event_extractor_logger.info.assert_called_once_with(events[0])